feat: refactor coding stats components to use centralized GitHub stats card and improve code organization
